src/google-cloud-speech/python/requirements.txt
google-cloud-speech >= 2.0
